Remove unused classes, restructure source tree
