+++ vaadinfrontend/src/main/java/life/qbic/datamanager/exceptionhandlers/CustomErrorHandler.java
@@ -10,8 +10,8 @@ import java.util.Locale;
 import life.qbic.application.commons.ApplicationException;
 import life.qbic.application.commons.ApplicationException.ErrorCode;
 import life.qbic.application.commons.ApplicationException.ErrorParameters;
-import life.qbic.datamanager.views.components.ErrorMessage;
-import life.qbic.datamanager.views.components.StyledNotification;
+import life.qbic.datamanager.views.notifications.ErrorMessage;
+import life.qbic.datamanager.views.notifications.StyledNotification;
 import life.qbic.logging.api.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
