+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/login/LoginHandler.java
@@ -6,8 +6,8 @@ import java.util.List;
 import java.util.Map;
 import life.qbic.authentication.application.user.registration.ConfirmEmailInput;
 import life.qbic.authentication.application.user.registration.ConfirmEmailOutput;
-import life.qbic.datamanager.views.components.ErrorMessage;
-import life.qbic.datamanager.views.components.InformationMessage;
+import life.qbic.datamanager.views.notifications.ErrorMessage;
+import life.qbic.datamanager.views.notifications.InformationMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
