+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/notifications/DisplayMessage.java
@@ -1,4 +1,4 @@
-package life.qbic.datamanager.views.components;
+package life.qbic.datamanager.views.notifications;
 
 import com.vaadin.flow.component.Composite;
 import com.vaadin.flow.component.html.Div;
